Use the correct type for the parameter to vpr::SystemPosix::Ntohs().
git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@21814 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vapor/vpr/md/POSIX/SystemPosix.h
@@ -116,7 +116,7 @@ public:
     * network byte ordering.  This is safe to use with signed and unsigned
     * values.
     */
-   static vpr::Uint16 Ntohs(vpr::Uint32 conversion)
+   static vpr::Uint16 Ntohs(vpr::Uint16 conversion)
    {
       return ntohs(conversion);
    }
